requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=14.0
google-cloud-bigquery>=3.18
google-auth>=2.26
scikit-learn>=1.4
matplotlib>=3.8
